fix(cli): register `service run` in the subcommand contract test
`cli::tests::service_subcommand_is_registered` pins the exact set of `service` sub-subcommands as a cross-platform contract, and the hidden `run` entry point added for the Windows Service Control Manager broke it. CI caught this; the scoped test runs used during implementation did not touch `cli::`.

The assertion now includes `run`, plus a new check that `run` is the only action hidden from `--help` and that every operator-facing action stays visible. The old assertion would have accepted a future change that hid `install`, which is exactly the kind of regression the contract exists to prevent.

Also splits the shutdown and readiness signalling out of `src/api/server.rs` into `src/api/shutdown.rs`. `server.rs` was already past the file-size soft limit before this branch and the new code pushed it to 769 lines; it is back to 648, and the new module is a self-contained concern (it is the only place that decides what counts as a reason to stop serving). The `api::shutdown` entry points are reached through the module path rather than a glob re-export, because a re-export is an unused import in the binary target, where the module tree is private.

Updates the `Commands::Service` help text, which still said the Windows backend was tracked separately.

Validation: `cargo fmt --check`, `cargo clippy --lib --tests -- -D warnings`, `cargo check --all-targets`, and `cargo test --lib` (1260 passed) on macOS. The Windows-only tree was re-checked with `cargo clippy --target x86_64-pc-windows-msvc --lib --bins --tests -- -D warnings` through the isolated probe crate, and probe coverage of each Windows-only file was re-confirmed by injecting a deliberate type error into each in turn.

Refs #311
